Move VIZ_TYPES to separate module
