+++ superset/assets/src/modules/vizTypes.js
@@ -1,0 +1,51 @@
+// You ***should*** use these to reference viz_types in code
+export const VIZ_TYPES = {
+  area: 'area',
+  bar: 'bar',
+  big_number: 'big_number',
+  big_number_total: 'big_number_total',
+  box_plot: 'box_plot',
+  bubble: 'bubble',
+  bullet: 'bullet',
+  cal_heatmap: 'cal_heatmap',
+  compare: 'compare',
+  directed_force: 'directed_force',
+  chord: 'chord',
+  dist_bar: 'dist_bar',
+  filter_box: 'filter_box',
+  heatmap: 'heatmap',
+  histogram: 'histogram',
+  horizon: 'horizon',
+  iframe: 'iframe',
+  line: 'line',
+  line_multi: 'line_multi',
+  mapbox: 'mapbox',
+  markup: 'markup',
+  para: 'para',
+  pie: 'pie',
+  pivot_table: 'pivot_table',
+  sankey: 'sankey',
+  separator: 'separator',
+  sunburst: 'sunburst',
+  table: 'table',
+  time_table: 'time_table',
+  time_pivot: 'time_pivot',
+  treemap: 'treemap',
+  country_map: 'country_map',
+  word_cloud: 'word_cloud',
+  world_map: 'world_map',
+  dual_line: 'dual_line',
+  event_flow: 'event_flow',
+  paired_ttest: 'paired_ttest',
+  partition: 'partition',
+  deck_scatter: 'deck_scatter',
+  deck_screengrid: 'deck_screengrid',
+  deck_grid: 'deck_grid',
+  deck_hex: 'deck_hex',
+  deck_path: 'deck_path',
+  deck_geojson: 'deck_geojson',
+  deck_multi: 'deck_multi',
+  deck_arc: 'deck_arc',
+  deck_polygon: 'deck_polygon',
+  rose: 'rose',
+};
