+++ superset/assets/src/visualizations/nvd3_vis.js
@@ -20,7 +20,7 @@ import { t } from '../locales';
 
 // CSS
 import './nvd3_vis.css';
-import { VIZ_TYPES } from './';
+import { VIZ_TYPES } from '../modules/vizTypes';
 
 const minBarWidth = 15;
 // Limit on how large axes margins can grow as the chart window is resized
